KAN-42 refactor: move logic out of SignInForm
