+++ src/features/auth/forms/SignInForm/SignInForm.tsx
@@ -10,37 +10,28 @@ import { z } from 'zod'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { Button } from '@/shared/components/Button'
 import Link from 'next/link'
-import { PrivateRoutes, PublicRoutes } from '@/shared/enums'
-import { useLoginMutation } from '@/features/auth/api'
-import { useRouter } from 'next/navigation'
+import { PublicRoutes } from '@/shared/enums'
 
 const signInSchema = z.object({
    email: z.email({ error: 'The email must match the format example@example.com' }),
    password: z.string({ error: 'The email or password are incorrect. Try again please' }),
 })
 
-type Inputs = z.infer<typeof signInSchema>
+export type Inputs = z.infer<typeof signInSchema>
 
-export const SignInForm = () => {
-   const [login] = useLoginMutation()
-   const router = useRouter()
+type Props = {
+   onSubmitAction: (data: Inputs) => void
+}
 
+export const SignInForm = ({ onSubmitAction }: Props) => {
    const {
       control,
       handleSubmit,
       formState: { errors },
    } = useForm<Inputs>({ resolver: zodResolver(signInSchema) })
 
-   const onSubmit: SubmitHandler<Inputs> = async data => {
-      const result = await login(data)
-      if (result.data) {
-         localStorage.setItem('accessToken', result.data.accessToken)
-         router.push(PrivateRoutes.feed)
-      }
-      // else if (result.error?.data?.errorsMessages?.length) {
-      //    //обработка ошибок
-      //    alert.error(result.error.data.errorsMessages[0].message)
-      // }
+   const onSubmit: SubmitHandler<Inputs> = data => {
+      onSubmitAction(data)
    }
    return (
       <Card className={'flex w-full max-w-[378px] flex-col items-center p-6'}>
